added changes untop of folder restructure
